backend/requirements.txt
adal==1.2.7
argcomplete==3.5.0
asgiref==3.8.1
attrs==24.2.0
azure-ai-ml==1.19.0
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-mgmt-core==1.4.0
azure-storage-blob==12.22.0
azure-storage-file-datalake==12.16.0
azure-storage-file-share==12.17.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.2.0
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.6.2
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docker==7.1.0
fonttools==4.53.1
frozendict==2.4.4
google-api-core==2.19.1
google-auth==2.33.0
googleapis-common-protos==1.63.2
html5lib==1.1
humanfriendly==10.0
idna==3.7
isodate==0.6.1
jeepney==0.8.0
jmespath==1.0.1
jsonpickle==3.2.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kaleido==0.2.1
kiwisolver==1.4.5
knack==0.11.0
lxml==5.2.2
marshmallow==3.21.3
matplotlib==3.9.2
msal==1.30.0
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4.post1
multitasking==0.0.11
ndg-httpsclient==0.5.1
numpy==2.0.1
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
packaging==24.1
pandas==2.2.2
paramiko==3.4.1
pathspec==0.12.1
peewee==3.17.6
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.2.2
plotly==5.23.0
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydash==8.0.3
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
#pywin32==306
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
schedule==1.2.2
SecretStorage==3.3.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
strictyaml==1.7.3
tabulate==0.9.0
tenacity==9.0.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
yfinance==0.2.41
